% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAanalysis-package.R
\docType{package}
\name{DNAanalysis-package}
\alias{DNAanalysis}
\alias{DNAanalysis-package}
\title{DNAanalysis: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
